Tardy-Recorder/requirements.txt
requests>=2.32
python-dotenv>=1.0.1
pydantic>=2.8
PyYAML>=6.0.2
openpyxl>=3.1.5
appdirs>=1.4.4
tk>=0.1.0